+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* VS Code companion: Non-Delphi source files are now correctly excluded from "Analyze All Open Files" analyses
+
 ## [0.5.0] - 2023-08-15
 
 ### Added